Seperated OrderBook impl and trait, working on inner methods
