+++ src/test/scala/s_mach/example/stock/OrderBookTest.scala
@@ -5,11 +5,12 @@ import org.scalatest.{Matchers, FlatSpec}
 class OrderBookTest extends FlatSpec with Matchers {
 
   "OrderBook" should "construct an order correctly" in {
-    val amzn = new OrderBook("AMZN")
+    val amzn = OrderBook("AMZN")
     assert(amzn.symbol == "AMZN")
-
+    val buyOrder = new BuyOrder("AMZN", 50, 5)
     //PQ ordering and priority
-
+    assert(amzn.addBuyer(buyOrder))
+    assert(amzn.addSeller(buyOrder)==false)
     //match testing
 
     //adding removing
